% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tablorg.R
\name{addConf}
\alias{addConf}
\title{Adding a configuration entry}
\usage{
addConf(conf, script = "", template = "", path = ".", run = TRUE)
}
\arguments{
\item{conf}{a \code{list}.}

\item{script}{a \code{string}. The filename including the path to the script
that will be executed based on the configuration.}

\item{template}{a \code{string}. A HTML template for the table. Either the
template as a text string or the path to a file containing the template.
Default value \code{""} means that no template is used.}

\item{path}{a \code{string}. The root directory of the computations where
the table will be stored.}

\item{run}{a \code{logical}. Should the script be run. Default value is
\code{TRUE}.}
}
\value{
an object of class 'tablorg'.
}
\description{
Adds a configuration entry to the table or creates a new table with the
configuration entry if no table exists.
}

