% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tablorg.R
\name{runEntry}
\alias{runEntry}
\title{Runs a tablorg entry script}
\usage{
runEntry(tabl, which, view = TRUE)
}
\arguments{
\item{tabl}{a \code{tablorg} object.}

\item{which}{an \code{integer}. Specifies which of the entries in the tablorg
object that will be executed. If missing, the active entry will be
executed.}

\item{view}{a \code{logical}. Should the results be}
}
\value{
URL for the path to the resulting output file (invisibly).
}
\description{
Given a tablorg object this function executes the script corresponding to
one entry in the table.
}

