% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tablorg-package.R
\docType{package}
\name{tablorg}
\alias{tablorg}
\alias{tablorg-package}
\title{tablorg: Table organization of computations}
\description{
Tools for organizing results from R computations. Overview and easy access
to the results are provided via an HTML table. This is a light weight alternative
to an actual database. The tools are useful when running many analyses or simulations
with varying parameter settings. The results are organized in directories
and accessed through an HTML table via a web server. This gives an R independent
way of monitoring computations running e.g. as an R batch script either
locally or on a remote server.
}
\author{
Niels Richard Hansen \email{Niels.R.Hansen@math.ku.dk}
}

