% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tablorg.R
\name{runEntry}
\alias{runEntry}
\title{Runs a tablorg entry script}
\usage{
runEntry(tabl, select, view = TRUE)
}
\arguments{
\item{tabl}{a \code{list} or a \code{string} specifying the path to a
json file.}

\item{select}{an \code{integer} or \code{character}. Specifies for which of the
entries in the table the correspondin script will be executed. If
missing, the active entry will be executed.}

\item{view}{a \code{logical}. Should the result be opened in a browser.}
}
\value{
URL for the path to the resulting output file (invisibly).
}
\description{
Given a tablorg object this function executes the script corresponding to
one entry in the table.
}

